# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "asp API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/mentor-grants": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "mentorUserId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "applicantUserId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "active",
                "revoked"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "mentorUserId": {
                        "type": "string"
                      },
                      "applicantUserId": {
                        "type": "string"
                      },
                      "permissions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "grantedByUserId": {
                        "type": "string"
                      },
                      "grantedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "status": {
                        "type": "string"
                      },
                      "requestedByUserId": {
                        "nullable": true,
                        "type": "string"
                      },
                      "applicantName": {
                        "type": "string"
                      },
                      "applicantEmail": {
                        "type": "string"
                      },
                      "mentorName": {
                        "type": "string"
                      },
                      "mentorEmail": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "mentorUserId",
                      "applicantUserId",
                      "permissions",
                      "grantedByUserId",
                      "grantedAt",
                      "status",
                      "applicantName",
                      "applicantEmail",
                      "mentorName",
                      "mentorEmail"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mentorUserId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "applicantUserId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "read",
                        "write"
                      ]
                    }
                  }
                },
                "required": [
                  "mentorUserId",
                  "applicantUserId",
                  "permissions"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "mentorUserId": {
                      "type": "string"
                    },
                    "applicantUserId": {
                      "type": "string"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "grantedByUserId": {
                      "type": "string"
                    },
                    "grantedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string"
                    },
                    "requestedByUserId": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "mentorUserId",
                    "applicantUserId",
                    "permissions",
                    "grantedByUserId",
                    "grantedAt",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/mentor-grants/mine": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "mentorUserId": {
                        "type": "string"
                      },
                      "applicantUserId": {
                        "type": "string"
                      },
                      "permissions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "grantedByUserId": {
                        "type": "string"
                      },
                      "grantedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "status": {
                        "type": "string"
                      },
                      "requestedByUserId": {
                        "nullable": true,
                        "type": "string"
                      },
                      "applicantName": {
                        "type": "string"
                      },
                      "applicantEmail": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "mentorUserId",
                      "applicantUserId",
                      "permissions",
                      "grantedByUserId",
                      "grantedAt",
                      "status",
                      "applicantName",
                      "applicantEmail"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/mentor-grants/my-requests": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "mentorUserId": {
                        "type": "string"
                      },
                      "applicantUserId": {
                        "type": "string"
                      },
                      "permissions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "grantedByUserId": {
                        "type": "string"
                      },
                      "grantedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "status": {
                        "type": "string"
                      },
                      "requestedByUserId": {
                        "nullable": true,
                        "type": "string"
                      },
                      "mentorName": {
                        "type": "string"
                      },
                      "mentorEmail": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "mentorUserId",
                      "applicantUserId",
                      "permissions",
                      "grantedByUserId",
                      "grantedAt",
                      "status",
                      "mentorName",
                      "mentorEmail"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/mentor-grants/requests": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mentorEmail": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "mentorEmail"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/mentor-grants/{id}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "revoked"
                    ]
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "read",
                        "write"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "mentorUserId": {
                      "type": "string"
                    },
                    "applicantUserId": {
                      "type": "string"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "grantedByUserId": {
                      "type": "string"
                    },
                    "grantedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string"
                    },
                    "requestedByUserId": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "mentorUserId",
                    "applicantUserId",
                    "permissions",
                    "grantedByUserId",
                    "grantedAt",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:6040"
    }
  ]
}